+++ test/token/ERC20/ERC20.behavior.js
@@ -88,7 +88,7 @@ function shouldBehaveLikeERC20 (errorPrefix, initialSupply, initialHolder, recip
 
             it('reverts', async function () {
               await expectRevert(this.token.transferFrom(
-                tokenOwner, to, amount, { from: spender }), 'SafeMath: subtraction overflow'
+                tokenOwner, to, amount, { from: spender }), `${errorPrefix}: transfer of tokens exceeding balance`
               );
             });
           });
@@ -104,7 +104,7 @@ function shouldBehaveLikeERC20 (errorPrefix, initialSupply, initialHolder, recip
 
             it('reverts', async function () {
               await expectRevert(this.token.transferFrom(
-                tokenOwner, to, amount, { from: spender }), 'SafeMath: subtraction overflow'
+                tokenOwner, to, amount, { from: spender }), `${errorPrefix}: transfer of tokens exceeding allowance`
               );
             });
           });
@@ -114,7 +114,7 @@ function shouldBehaveLikeERC20 (errorPrefix, initialSupply, initialHolder, recip
 
             it('reverts', async function () {
               await expectRevert(this.token.transferFrom(
-                tokenOwner, to, amount, { from: spender }), 'SafeMath: subtraction overflow'
+                tokenOwner, to, amount, { from: spender }), `${errorPrefix}: transfer of tokens exceeding balance`
               );
             });
           });
@@ -166,7 +166,7 @@ function shouldBehaveLikeERC20Transfer (errorPrefix, from, to, balance, transfer
 
       it('reverts', async function () {
         await expectRevert(transfer.call(this, from, to, amount),
-          'SafeMath: subtraction overflow'
+          `${errorPrefix}: transfer of tokens exceeding balance`
         );
       });
     });